Create a minimum base framework
The intention is to construct a framework taking inspiration from what
works well for web-frameworks such as Flask and Litestar.

While I'm unsure whether the use of `ContextVar` for application context
is the best choice or if a dependency injection approach akin to that of
Litestar would be better, we'll stick for this for now and transition if
we identify the need.