Add a caching helper
These cache layers can be used by users of the framework to cache
results of expensive function invocations. Consider a check that has to
interact with a third-party service's API and gather a lot of data which
might be both time-consuming and maybe expensive. The user can then
define an in-memory or a persistent disk-cache in which the result of
the relevant function can be memoized to avoid repeated calls.

This exists for two reasons:

* The cache supports a TTL, which the default functools caches don't
  support.
* The disk-based cache is extremely helpful when a function has to work
  with large files.

  One real-life example is a check that verifies the sitemaps of a
  website, where the sum of the XMLs is multiple hundreds of megabytes.
  Using an on-disk cache and smart HEAD requests, the check that
  verifies the sitemaps can then avoid downloading sitemaps that have
  already been downloaded and haven't changed.