Ensure tests don't depend on check execution speed
With the check executor now being integrated, there is a time gap
between when a check is submitted and when it is ready. In my test
executions so far the gap between submission and readiness was shorter
than the gap between submission and _checking_ for readiness -- in other
words, the tests always succeeded, not encountering any in-flight
checks.

To make sure this isn't luck, though, this commit allows the outpost app
to be launched with a custom executor, which the tests make use of by
providing a `BlockingCheckExecutor` which ensures any requested future
has completed before returning its result.