Rename strategy from exec-environment to target environment
The name "MustRunInCurrentExecutionEnvironment" is a tautology, in that
if a check is running, it is inevitably running in the current execution
environment. What this strategy is actually saying is that the execution
environment the check can run in must be the target environment the
check is running for, hence the name change to
"MustRunInTargetEnvironment" to better describe this.