Support combining Datasource and DatasourceFactory
Prior to this commit we always instantiated a registered
`DatasourceFactory`. This was never necessary because we did not allow
to provide any arguments to the factory itself. Now we no longer
instantiate it, which allows the `new` method to be a classmethod (or
even a staticmethod).

The result of this is that a class implementing `Datasource` can now
just as easily also fulfill the `DatasourceFactory` protocol,
alleviating the need for separate classes.

This also makes defining the `FromFactory` annotation simpler, because
we no longer require the factory-type to be specified. If it wasn't
specified, we assume that the type that was annotated is also the
factory.

Closes #35.