Rename Outpost to Watchpost
While I liked the outpost name, it had one fatal flaw: there is already
a package with that name on pypi that has no relation to this project.
While there is currently no intention to publish this watchpost project
to pypi, users that discover it might attempt to `pip install outpost`
and be confused what it is they are installing.