Delay datasource instantiation until check invocation
Certain datasources might not be instantiable in some environments, and
that is not only OK, but something we explicitly want to support.

What we intended to implement is that datasources used by checks that
aren't scheduled are not attempted to be instantiated at all.
Unfortunately, the way the scheduling was implemented, the datasources
were instantiated to find out.

This commit addresses this problem, delaying the instantiation right up
until the check is actually invoked.

Closes #41.